refactor(lib): migrate session utility to TypeScript and update imports
