+++ apps/manage/src/app/views/cases/view/manage-reps/add/save.js
@@ -2,7 +2,7 @@ import { uniqueReference } from '@pins/crowndev-lib/util/random-reference.js';
 import { notFoundHandler } from '@pins/crowndev-lib/middleware/errors.js';
 import { isValidUuidFormat } from '@pins/crowndev-lib/util/uuid.ts';
 import { JOURNEY_ID } from './journey.js';
-import { addSessionData, clearSessionData, readSessionData } from '@pins/crowndev-lib/util/session.js';
+import { addSessionData, clearSessionData, readSessionData } from '@pins/crowndev-lib/util/session.ts';
 import { saveRepresentation } from '@pins/crowndev-lib/forms/representations/save.js';
 import {
 	deleteRepresentationAttachmentsFolder,
