+++ apps/manage/src/app/views/cases/view/manage-reps/edit/controller.js
@@ -1,7 +1,7 @@
 import { editsToDatabaseUpdates } from '@pins/crowndev-lib/forms/representations/view-model.js';
 import { wrapPrismaError } from '@pins/crowndev-lib/util/database.js';
 import { validateParams } from '../view/controller.js';
-import { addSessionData, clearSessionData, readSessionData } from '@pins/crowndev-lib/util/session.js';
+import { addSessionData, clearSessionData, readSessionData } from '@pins/crowndev-lib/util/session.ts';
 import {
 	publishedRepresentationsAttachmentsRootFolderPath,
 	representationFolderPath
