+++ apps/manage/src/app/views/cases/view/manage-reps/review/controller.js
@@ -1,7 +1,7 @@
 import { REPRESENTATION_STATUS_ID } from '@pins/crowndev-database/src/seed/data-static.ts';
 import { ACCEPT_AND_REDACT } from '@pins/crowndev-lib/forms/representations/questions.js';
 import { renderRepresentation, validateParams } from '../view/controller.js';
-import { addSessionData, clearSessionData, readSessionData } from '@pins/crowndev-lib/util/session.js';
+import { addSessionData, clearSessionData, readSessionData } from '@pins/crowndev-lib/util/session.ts';
 import { createRedactJourney } from './journey.js';
 import { JOURNEY_ID } from '../view/journey.js';
 import { JourneyResponse } from '@planning-inspectorate/dynamic-forms/src/journey/journey-response.js';
