+++ apps/manage/src/app/views/cases/view/manage-reps/validation-middleware.js
@@ -1,6 +1,6 @@
 import { REPRESENTATION_SUBMITTED_FOR_ID, REPRESENTED_TYPE_ID } from '@pins/crowndev-database/src/seed/data-static.ts';
 import { notFoundHandler } from '@pins/crowndev-lib/middleware/errors.js';
-import { addSessionData } from '@pins/crowndev-lib/util/session.js';
+import { addSessionData } from '@pins/crowndev-lib/util/session.ts';
 import { fileAlreadyExistsInFolder } from '@pins/crowndev-lib/forms/custom-components/representation-attachments/document-validation-util.js';
 import { expressValidationErrorsToGovUkErrorList } from '@planning-inspectorate/dynamic-forms/src/validator/validation-error-handler.js';
 import { buildReviewControllers } from './review/controller.js';
