+++ apps/portal/src/app/views/applications/view/have-your-say/controller.js
@@ -3,7 +3,7 @@ import { notFoundHandler } from '@pins/crowndev-lib/middleware/errors.js';
 import { applicationLinks } from '../view-model.ts';
 import { fetchPublishedApplication } from '#util/applications.ts';
 import { nowIsWithinRange } from '@planning-inspectorate/dynamic-forms';
-import { clearSessionData, readSessionData } from '@pins/crowndev-lib/util/session.js';
+import { clearSessionData, readSessionData } from '@pins/crowndev-lib/util/session.ts';
 import { loadPublishedApplicationOr404, shouldDisplayApplicationUpdatesLink } from '../../../util/application-util.ts';
 import {
 	buildRequiredCheckboxGroup,
