+++ apps/portal/src/app/views/applications/view/have-your-say/save.js
@@ -3,7 +3,7 @@ import { notFoundHandler } from '@pins/crowndev-lib/middleware/errors.js';
 import { crownDevelopmentToViewModel } from '../view-model.ts';
 import { fetchPublishedApplication } from '#util/applications.ts';
 import { formatDateForDisplay } from '@planning-inspectorate/dynamic-forms';
-import { addSessionData, clearSessionData, readSessionData } from '@pins/crowndev-lib/util/session.js';
+import { addSessionData, clearSessionData, readSessionData } from '@pins/crowndev-lib/util/session.ts';
 import { uniqueReference } from '@pins/crowndev-lib/util/random-reference.js';
 import { JOURNEY_ID } from './journey.js';
 import { REPRESENTATION_SUBMITTED_FOR_ID } from '@pins/crowndev-database/src/seed/data-static.ts';
