+++ packages/lib/forms/custom-components/representation-attachments/upload-documents.js
@@ -4,7 +4,7 @@ import { FILE_PROPERTIES } from '../../../documents/view-model.js';
 import { isValidUuidFormat } from '../../../util/uuid.ts';
 import { notFoundHandler } from '../../../middleware/errors.js';
 import { sortByField } from '../../../util/array.ts';
-import { addSessionData } from '../../../util/session.js';
+import { addSessionData } from '../../../util/session.ts';
 import { getSubmittedForId } from '../../../util/questions.js';
 import { getApplicationNameFolder } from '../../../util/handle-attachments.js';
