+++ packages/lib/forms/representations/save.js
@@ -1,4 +1,4 @@
-import { addSessionData, clearSessionData } from '../../util/session.js';
+import { addSessionData, clearSessionData } from '../../util/session.ts';
 import { viewModelToRepresentationCreateInput } from './view-model.js';
 import { clearDataFromSession } from '@planning-inspectorate/dynamic-forms/src/lib/session-answer-store.js';
 import { wrapPrismaError } from '../../util/database.js';
